Move validateCache logic outside SplitsCacheInLocal
