+++ src/storages/inLocalStorage/__tests__/SplitsCacheInLocal.spec.ts
@@ -7,7 +7,6 @@ import { fullSettings } from '../../../utils/settingsValidation/__tests__/settin
 
 test('SPLIT CACHE / LocalStorage', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   cache.clear();
 
@@ -41,7 +40,6 @@ test('SPLIT CACHE / LocalStorage', () => {
 
 test('SPLIT CACHE / LocalStorage / Get Keys', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   cache.addSplit('lol1', something);
   cache.addSplit('lol2', somethingElse);
@@ -54,7 +52,6 @@ test('SPLIT CACHE / LocalStorage / Get Keys', () => {
 
 test('SPLIT CACHE / LocalStorage / Add Splits', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   cache.addSplits([
     ['lol1', something],
@@ -69,7 +66,6 @@ test('SPLIT CACHE / LocalStorage / Add Splits', () => {
 
 test('SPLIT CACHE / LocalStorage / trafficTypeExists and ttcache tests', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   cache.addSplits([ // loop of addSplit
     ['split1', splitWithUserTT],
@@ -108,7 +104,6 @@ test('SPLIT CACHE / LocalStorage / trafficTypeExists and ttcache tests', () => {
 
 test('SPLIT CACHE / LocalStorage / killLocally', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   cache.addSplit('lol1', something);
   cache.addSplit('lol2', somethingElse);
@@ -142,7 +137,6 @@ test('SPLIT CACHE / LocalStorage / killLocally', () => {
 
 test('SPLIT CACHE / LocalStorage / usesSegments', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   expect(cache.usesSegments()).toBe(true); // true initially, until data is synchronized
   cache.setChangeNumber(1); // to indicate that data has been synced.
@@ -174,7 +168,6 @@ test('SPLIT CACHE / LocalStorage / flag set cache tests', () => {
       }
     }
   }, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   const emptySet = new Set([]);
 
@@ -216,7 +209,6 @@ test('SPLIT CACHE / LocalStorage / flag set cache tests', () => {
 // if FlagSets are not defined, it should store all FlagSets in memory.
 test('SPLIT CACHE / LocalStorage / flag set cache tests without filters', () => {
   const cache = new SplitsCacheInLocal(fullSettings, new KeyBuilderCS('SPLITIO', 'user'));
-  cache.validateCache(fullSettings);
 
   const emptySet = new Set([]);
 
@@ -236,6 +228,5 @@ test('SPLIT CACHE / LocalStorage / flag set cache tests without filters', () => 
 
   // Validate that the feature flag cache is cleared when calling `clear` method
   cache.clear();
-  expect(localStorage.length).toBe(1); // only 'SPLITIO.hash' should remain in localStorage
-  expect(localStorage.key(0)).toBe('SPLITIO.hash');
+  expect(localStorage.length).toBe(0);
 });
